// artifacts/api-server/src/sya/lib/schemas.ts
export const ListCasesQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  year: zod.coerce.string().optional(),
});
// ...
export const GetCaseParams = zod.object({
  id: zod.coerce.number(),
});